requirements.txt
Django==2.1
gunicorn==19.9.0
psycopg2==2.7
redis==2.10.3
djangorestframework==3.8.2